requirements.txt
# Root requirements.txt – used by Vercel's Python serverless runtime.
# Only includes deps needed for the serverless function (psycopg, not asyncpg).
fastapi>=0.115.0
sqlalchemy[asyncio]>=2.0.36
psycopg[binary]>=3.1.0
python-jose[cryptography]>=3.3.0
google-auth>=2.37.0
requests>=2.32.0
pydantic>=2.10.0
